+++ packages/firebase_auth/firebase_auth_platform_interface/test/method_channel_firebase_auth_test.dart
@@ -209,6 +209,26 @@ void main() {
       );
     });
 
+    test('signInAnonymously with null additionalUserInfo', () async {
+      MethodChannelFirebaseAuth.channel
+          .setMockMethodCallHandler((MethodCall call) async {
+        log.add(call);
+        return <String, dynamic>{
+          'user': kMockUser,
+        };
+      });
+      final PlatformAuthResult result = await auth.signInAnonymously(appName);
+      verifyUser(result.user);
+      expect(result.additionalUserInfo, isNull);
+      expect(
+        log,
+        <Matcher>[
+          isMethodCall('signInAnonymously',
+              arguments: <String, String>{'app': appName}),
+        ],
+      );
+    });
+
     test('sendLinkToEmail', () async {
       await auth.sendLinkToEmail(
         appName,